Update to prop-types v15.5.10
React prop-types v15.5.6 has multiple issues (especially in production
minified builds), and was recently depreciated with a suggestion to
upgrade to v15.5.7 or higher. v15.5.10 is the most recent release, does
not introduce any breaking changes, and seems to work perfectly fine
with this package. By doing this we can get rid of a scary depreciation
message whenever this package is installed :)